# packages/contracts/tsp-output/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: (title)
  description: |-
    Project API - RESTful API following modern best practices

    Key Features:
    - URL versioning (/v1)
    - Multi-tenant architecture (path-based with /orgs/{orgId})
    - Comprehensive pagination with navigation links
    - Structured error responses with detailed validation
    - Soft delete support with restore capability
    - Batch operations for bulk processing
    - Field selection and relation inclusion
    - Full-text search and filtering
  version: 0.0.0
...
paths:
...
  /v1/orgs/{orgId}/users/{userId}/roles:
    get:
      operationId: UserTenantRoles_list
      summary: List user roles in tenant
      description: |-
        List roles assigned to a user in this tenant

        Returns both global and tenant-specific role assignments.
      parameters:
        - name: orgId
          in: path
          required: true
          description: Organization ID
          schema:
            type: string
        - name: userId
          in: path
          required: true
          description: User ID
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/UserRoleAssignmentListResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Roles
    post:
      operationId: UserTenantRoles_assign
      summary: Assign role to user
      description: Assign a role to a user in this tenant
      parameters:
        - name: orgId
          in: path
          required: true
          description: Organization ID
          schema:
            type: string
        - name: userId
          in: path
          required: true
          description: User ID
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRoleAssignmentResponse'
      tags:
        - Roles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRoleRequest'
        description: Role assignment
  /v1/orgs/{orgId}/users/{userId}/roles/{roleId}:
    delete:
      operationId: UserTenantRoles_remove
      summary: Remove role from user
      description: Remove a role from a user
      parameters:
        - name: orgId
          in: path
          required: true
          description: Organization ID
          schema:
            type: string
        - name: userId
          in: path
          required: true
          description: User ID
          schema:
            type: string
        - name: roleId
          in: path
          required: true
          description: Role ID to remove
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
      tags:
        - Roles
...
components:
  schemas:
...
    AssignRoleRequest:
      type: object
      required:
        - roleId
      properties:
        roleId:
          type: string
          description: Role ID to assign
      description: Request to assign a role to a user
...
    Error:
      type: object
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
          description: Array of detailed error information (e.g., validation errors)
        requestId:
          type: string
          description: Request ID for tracking and debugging
        documentationUrl:
          type: string
          description: URL to documentation about this error (optional)
      description: Standard error object
    ErrorDetail:
      type: object
      required:
        - field
        - code
        - message
      properties:
        field:
          type: string
          description: The field that caused the error (for validation errors)
        code:
          type: string
          description: Specific error code for this detail
        message:
          type: string
          description: Human-readable error message
        metadata:
          type: object
          additionalProperties: {}
          description: Additional metadata about the error
      description: Detailed error information for a specific field or validation
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          allOf:
            - $ref: '#/components/schemas/Error'
          description: The error object
      description: Standard error response envelope
...
    ResponseMeta:
      type: object
      required:
        - requestId
        - timestamp
      properties:
        requestId:
          type: string
          description: Unique request identifier for tracking and debugging
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the response was generated
        apiVersion:
          type: string
          description: API version used for this request
        tenantId:
          type: string
          description: Tenant/organization ID for multi-tenant requests
        tenantName:
          type: string
          description: Tenant/organization name for multi-tenant requests
        durationMs:
          type: integer
          format: int32
          description: Processing time in milliseconds
      description: Response metadata included in all API responses
...
    UserRoleAssignment:
      type: object
      required:
        - id
        - userId
        - applicationId
        - roleId
        - roleName
        - isGlobalRole
        - assignedAt
      properties:
        id:
          type: string
          description: Assignment ID
          example: ura_abc123
        userId:
          type: string
          description: User ID
          example: usr_123
        applicationId:
          type: string
          description: Application ID
          example: app_default
        tenantId:
          type: string
          description: Tenant ID (null for global role assignments)
          example: org_abc123
        roleId:
          type: string
          description: Role ID
          example: role_admin
        roleName:
          type: string
          description: Role name (denormalized for convenience)
          example: Admin
        isGlobalRole:
          type: boolean
          description: Whether this is a global role assignment
        assignedAt:
          type: string
          format: date-time
          description: When the role was assigned
        assignedBy:
          type: string
          description: Who assigned the role
          example: usr_owner
      description: |-
        User's role assignment

        Represents a single role assigned to a user within a specific
        application and optional tenant context.
    UserRoleAssignmentListResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/UserRoleAssignment'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
      description: User role assignments list response
    UserRoleAssignmentResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/UserRoleAssignment'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
      description: User role assignment response
...
servers:
  - url: https://api.example.com
    description: Production server
    variables: {}
  - url: http://localhost:3001
    description: Development server
    variables: {}